requirements.txt
Flask
Werkzeug
opencv-python
matplotlib
numpy
Flask-Cors
PyJWT
google-genai
PyPDF2
python-docx
